refactor : money 값 double로 개선

+++ src/main/java/lotto/domain/lotto/LottoMachine.java
@@ -13,7 +13,7 @@ class LottoMachine {
     private static final Money LOTTO_PRICE = Money.from(1000);
     private static final String UNKNOWN_FACTORY_MESSAGE = "알 수 없는 로또 생성기로 로또 기계를 생성할 수 없습니다.";
     private static final String INVALID_MONEY_MESSAGE =
-            "금액을 " + LOTTO_PRICE.getValue() + Money.CURRENCY + " 단위로 입력해주세요.";
+            "금액을 " + (int) LOTTO_PRICE.getValue() + Money.CURRENCY + " 단위로 입력해주세요.";
 
     private final LottoFactory factory;
 
@@ -36,7 +36,7 @@ class LottoMachine {
     public List<Lotto> issueWith(Money money) {
         checkHasRemainderWith(money);
 
-        return issue(money.divide(LOTTO_PRICE));
+        return issue((int) money.divide(LOTTO_PRICE));
     }
 
     private void checkHasRemainderWith(Money money) {
